WIP nix packaging
This is partial progress of packaging for https://nixos.org/nix/ and used node2nix 1.6.0 to generate the .nix-files.

The supplement feature and `node2nix -l` can't be used together so I removed the package-lock.json.

As posted in %Vx/1OSDLnXXu4P3Cmo4Mq8ZTO2kRRpT32o9JAUoFG8o=.sha256, It currently fails to use the correct location of the systray helper.